api/requirements.txt
fastapi==0.126.0
uvicorn[standard]==0.38.0
python-multipart==0.0.21
pydantic==2.12.5
qdrant-client==1.14.0
openai==1.14.0
python-dotenv==1.0.1
slowapi==0.1.9